# docs/openapi.json
{"openapi":"3.0.1","info":{"title":"Weather API","description":"Weather API"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Authentication","description":"Authentication and registration users"},{"name":"Forecasts"},{"name":"Locations"}],"paths":{"/api/location/save":{"post":{"tags":["Locations"],"summary":"Save location","description":"Save location for the user","operationId":"saveLocation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationApiResponseDto"}}},"required":true},"responses":{"204":{"description":"No Content"},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorDto"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"}},"security":[{"bearer":[]}]}},"/api/location/delete":{"post":{"tags":["Locations"],"summary":"Delete location","description":"Delete location from user with 'id'","operationId":"deleteLocation","requestBody":{"content":{"application/json":{"schema":{"type":"integer","format":"int64"}}},"required":true},"responses":{"204":{"description":"No Content"},"409":{"description":"Conflict","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorDto"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorDto"}}}},"404":{"description":"Not Found"}},"security":[{"bearer":[]}]}},"/api/forecast":{"post":{"tags":["Forecasts"],"summary":"Get list of forecasts","description":"Get forecasts from external API server","operationId":"getWeatherForecast","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocationReadDto"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ForecastDto"}}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"}},"security":[{"bearer":[]}]}},"/api/auth/registration":{"post":{"tags":["Authentication"],"summary":"Registration for users","description":"Registration with login and password","operationId":"registration","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateDto"}}},"required":true},"responses":{"201":{"description":"Success registration of user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserReadDto"}}}},"400":{"description":"Bad request","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApiErrorDto"}}}}},"409":{"description":"User with this login already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorDto"}}}}}}},"/api/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh token","description":"Refresh an access token and a refresh token with a validated refresh token","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}},"required":true},"responses":{"200":{"description":"Success refresh tokens","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtResponseDto"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login for users","description":"Required login and password","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Return access and refresh tokens","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtResponseDto"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"}}}},"/api/locations":{"get":{"tags":["Locations"],"summary":"Get user's locations","description":"Get slice of locations for user","operationId":"getLocations","parameters":[{"name":"pageNumber","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SliceOfLocations"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"}},"security":[{"bearer":[]}]}},"/api/locations/search":{"get":{"tags":["Locations"],"summary":"Get locations API server","description":"Get list of locations from external API server","operationId":"searchLocationsByName","parameters":[{"name":"locationName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocationApiResponseDto"}}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"}},"security":[{"bearer":[]}]}}},"components":{"schemas":{"UserReadDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"login":{"type":"string"}}},"ApiErrorDto":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"},"errorCode":{"type":"string"}}},"LocationApiResponseDto":{"type":"object","properties":{"name":{"type":"string"},"lat":{"type":"number","format":"double"},"lon":{"type":"number","format":"double"},"country":{"type":"string"},"state":{"type":"string"}}},"ForecastDto":{"type":"object","properties":{"locationId":{"type":"integer","format":"int64"},"locationName":{"type":"string"},"countryCode":{"type":"string"},"state":{"type":"string"},"iconUrl":{"type":"string"},"description":{"type":"string"},"temperature":{"type":"number","format":"float"},"pressure":{"type":"integer","format":"int32"},"humidity":{"type":"integer","format":"int32"}}},"LocationReadDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"state":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"UserCreateDto":{"required":["login","rawPassword"],"type":"object","properties":{"login":{"maxLength":20,"minLength":3,"type":"string"},"rawPassword":{"type":"string"},"confirmRawPassword":{"type":"string"}}},"JwtResponseDto":{"type":"object","properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"}}},"RefreshTokenDto":{"required":["refresh_token"],"type":"object","properties":{"refresh_token":{"type":"string"}}},"User":{"required":["login","password"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"login":{"type":"string"},"password":{"type":"string"}}},"SliceOfLocations":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/LocationReadDto"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"hasNext":{"type":"boolean"}}}},"responses":{"BadRequest":{"description":"Bad request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorDto"}}}},"Unauthorized":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiErrorDto"}}}}},"securitySchemes":{"bearer":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}